car_rent/requirements.txt
asgiref==3.5.2
branca==0.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
dnspython==2.4.2
folium==0.14.0
future==0.18.3
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
oauthlib==3.2.2
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.5.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==65.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.12
